🧪 :: 기능별 핵심 테스트로 정리
프로젝트 규모에 비해 세분화되어 있던 위젯/유틸 테스트를 기능 단위 테스트 파일로 합치고, 날짜 포맷/예약 상태/고장 신고 흐름처럼 비즈니스 로직 검증 중심으로 재구성했다.

Confidence: high

Scope-risk: narrow

Tested: flutter test
